@alamb 2025-07-03
🤖 `./gh_compare_arrow.sh` [Benchmark Script](https://github.com/alamb/datafusion-benchmarking/blob/main/gh_compare_arrow.sh) Running
Linux aal-dev 6.11.0-1016-gcp #16~24.04.1-Ubuntu SMP Wed May 28 02:40:52 UTC 2025 x86_64 x86_64 x86_64 GNU/Linux
Comparing pushdown-v4 (1851f0b6345a58d904e4a18e3f9b0b09fbf99b8e) to af8564f076109e79329063fc6c8fbb672e35c32e [diff](https://github.com/apache/arrow-rs/compare/af8564f076109e79329063fc6c8fbb672e35c32e..1851f0b6345a58d904e4a18e3f9b0b09fbf99b8e)
BENCH_NAME=arrow_reader_clickbench
BENCH_COMMAND=cargo bench --features=arrow,async,test_common,experimental --bench arrow_reader_clickbench 
BENCH_FILTER=
BENCH_BRANCH_NAME=pushdown-v4
Results will be posted here when complete