@alamb 2025-07-03
> 🤖: Benchmark completed

😎  -- very nice