@alamb 2025-07-03
> I am curious about the performance compared with no filter pushdown case, because previous try will also improve the performance for this benchmark. But compared to the no filter pushdown case, it has some regression.

I will try and run this experiment later today